Bundle the Ticker class with ShiftBrite
While the Ticker class technically isn't related to *driving hardware*
for ShiftBrite chains, I don't view the ShiftBrite class on its own as
being terribly useful for an actual product. The Ticker class allows us
to actually control the chain in a meaningful way.

Although come to think of it, the Ticker is hardware agnostic...it could
manage *any* RGB pixels, not just ShiftBrites, so maybe it warrants a
library of its own?